+++ demos/manual/src/main/java/is/codion/manual/design/DesignAndStyle.java
@@ -27,7 +27,7 @@ import is.codion.common.value.ValueList;
 import is.codion.framework.db.local.LocalEntityConnection;
 import is.codion.framework.db.local.LocalEntityConnectionProvider;
 import is.codion.framework.domain.entity.Entity;
-import is.codion.framework.model.EntityConditionModel;
+import is.codion.framework.model.EntityTableConditionModel;
 import is.codion.manual.store.domain.Store.Customer;
 import is.codion.swing.common.model.component.table.FilterTableModel;
 import is.codion.swing.common.model.component.table.FilterTableModel.TableSelection;
@@ -39,6 +39,7 @@ import java.util.List;
 import java.util.concurrent.TimeUnit;
 
 import static is.codion.common.event.Event.event;
+import static is.codion.framework.model.EntityTableConditionModel.entityTableConditionModel;
 import static javax.swing.BorderFactory.createTitledBorder;
 
 public final class DesignAndStyle {
@@ -56,8 +57,8 @@ public final class DesignAndStyle {
 
 		State state = State.state(true);
 
-		EntityConditionModel conditionModel =
-						EntityConditionModel.entityConditionModel(Customer.TYPE, connectionProvider);
+		EntityTableConditionModel conditionModel =
+						entityTableConditionModel(Customer.TYPE, connectionProvider);
 		//end::factories[]
 
 		//tag::builders[]